% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{defender}
\alias{defender}
\title{Convenience function to create a Defense pitch_object dataframe which tracks another objects movements.}
\usage{
defender(
  player,
  label = "",
  show = NA,
  object = "Defense",
  xjust = -1,
  yjust = -1,
  frame = NA,
  alpha = NA,
  label_pos = NA
)
}
\arguments{
\item{player}{Name of the object_dataframe to track (required)}

\item{label}{Name for the object (required)}

\item{show}{Show the label in plots : default = Same as player object, T, F}

\item{object}{Name for the type of object : "Defense" (default), "Offense", "Coach", "Cone", "Disc" are recognised but can be anything.}

\item{xjust}{Vector of adjustments to player x-coords to offset defenders position : default = -1}

\item{yjust}{Vector of adjustments to player y-coords to offset defenders position : default = -1}

\item{frame}{Vector of frames to show the object in : default = same as player object}

\item{alpha}{Colour opacity of the object : default = same as player object}

\item{label_pos}{Position of label relative to object : default = same as player object, "down"  "up", "left", "right"}
}
\value{
A data.frame
}
\description{
Convenience function to create a Defense pitch_object dataframe which tracks another objects movements.
}
\examples{
defender(player_list[[1]], "D1", x = c(-1,-1,5), y = c(2,2,5))
}
