% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{disc}
\alias{disc}
\title{Convenience function to create a pitch_object which follows the path of an arrow. E.g. a disc following a throw.}
\usage{
disc(
  throw,
  label,
  show = F,
  xjust = c(-1, -1),
  yjust = c(1, 1),
  frame = 1,
  release_frame = 2,
  alpha = 1,
  label_pos = "down"
)
}
\arguments{
\item{throw}{A pitch_arrow dataframe (required)}

\item{label}{Name for the object (required)}

\item{show}{Show the label in plots : default = F}

\item{xjust}{2 value vector showing the x-coord offset for the thrower and receiver : default = c(-1,-1)}

\item{yjust}{2 value vector showing the y-coord offset for the thrower and receiver : default = c(1,1)}

\item{frame}{Vector of frames to show the object in : default = 1}

\item{release_frame}{The frame where the disc should be released from the thrower : default = 2}

\item{alpha}{Colour opacity of the object : default = 1}

\item{label_pos}{Position of label relative to start or end of object : "down" (default), "up", "left", "right"}
}
\value{
A data.frame
}
\description{
Convenience function to create a pitch_object which follows the path of an arrow. E.g. a disc following a throw.
}
\examples{
disc(throw_list[[1]],"D1",frame = 1:3, release_frame = 2)
}
