% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{export_frames}
\alias{export_frames}
\title{Export an object or arrow list as a tsv file arranged by frames. This is useful for making manual edits.}
\usage{
export_frames(list, filename = "frames.tsv")
}
\arguments{
\item{list}{Either a list of objects or arrows}

\item{filename}{Name of the file to save : default = "frames.tsv"}
}
\value{
A data.frame
}
\description{
Export an object or arrow list as a tsv file arranged by frames. This is useful for making manual edits.
}
\examples{
export_frames(object_list)
}
