% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{ggpitch}
\alias{ggpitch}
\title{Generate a pitch map as a ggplot object. The 'type' parameter controls the section of pitch.
Default size is set to 100 x 37, with endzones 18 deep.}
\usage{
ggpitch(
  type = "full",
  colour = "dimgray",
  fill = "white",
  endzone_fill = "#AA4422",
  linewidth = 0.5,
  alpha = 1,
  endzone_alpha = 0.2,
  linetype = "solid",
  xmin = 0,
  xmax = 37,
  ymin = 0,
  ymax = 100,
  zone_depth = 18
)
}
\arguments{
\item{type}{Style of pitch: "full" (default), "half_attack", "half_defend", "blank", "void"}

\item{colour}{Colour of pitch outline : default = "dimgrey"}

\item{fill}{Fill colour of central zone  : default = "white"}

\item{endzone_fill}{Fill colour of endzone : default = #AA4422}

\item{linewidth}{Width of lines : default = 0.5}

\item{alpha}{Fill opacity of central zone : default = 1}

\item{endzone_alpha}{Fill opacity of endzone : default = 0.2}

\item{linetype}{Linetype of lines : default = solid}

\item{xmin}{Start x coordinate : default = 0}

\item{xmax}{End x coordinate : default = 37}

\item{ymin}{Start y coordinate : default = 0}

\item{ymax}{End y coordinate : default = 100}

\item{zone_depth}{Depth of endzones : default = 18}
}
\value{
A ggplot object
}
\description{
Generate a pitch map as a ggplot object. The 'type' parameter controls the section of pitch.
Default size is set to 100 x 37, with endzones 18 deep.
}
\examples{
ggpitch()
ggpitch("half_attack",endzone_fill="blue")
}
