% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{hstack}
\alias{hstack}
\title{Helper function to print a list of 7 players in horizontal stack formation}
\usage{
hstack()
}
\value{
A tibble
}
\description{
Helper function to print a list of 7 players in horizontal stack formation
}
\examples{
hstack
}
