% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{object_paths}
\alias{object_paths}
\title{Convenience function to create pitch_arrows which trace an objects movement path on the pitch e.g. a players cut. You can choose
to only trace specific object types e.g offense players only.}
\usage{
object_paths(object_list, objects = c("Offense"))
}
\arguments{
\item{object_list}{A list of object data.frames e.g. generated by pitch_object (required)}

\item{objects}{A vector of the object types to trace : default = "Offense"}
}
\value{
A data.frame
}
\description{
Convenience function to create pitch_arrows which trace an objects movement path on the pitch e.g. a players cut. You can choose
to only trace specific object types e.g offense players only.
}
\examples{
object_paths(player_list)
object_paths(player_list, c("Offense","Defense"))
}
