% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{pitch_arrow}
\alias{pitch_arrow}
\title{Create a pitch_arrow dataframe. Arrows are represented with geom_segment or geom_curve in the plot_play() function.
Arrows are usually player movements (cuts) or throws on the field.}
\usage{
pitch_arrow(
  label = "",
  show = F,
  object = "Offense",
  type = "Cut",
  x,
  y,
  xend,
  yend,
  frame = 1,
  alpha = 1,
  arrow_shape = "straight",
  label_pos = "start_down"
)
}
\arguments{
\item{label}{Name for the arrow (required)}

\item{show}{Show the label in plots : default = F}

\item{object}{Attribute an arrow to an object type to match colours : "Offense" (default), "Defense", "Coach", "Cone", "Disc" are recognised but can be anything.}

\item{type}{What the arrow represents : "Cut" (default), "Throw", "Label" are recognised but can be anything}

\item{x}{x-coord for start of arrow (required)}

\item{y}{y-coord for start of arrow (required)}

\item{xend}{x-coord for end of arrow (required)}

\item{yend}{y-coord for end of arrow (required)}

\item{frame}{Vector of frames to show the arrow in : default = 1}

\item{alpha}{Colour opacity of the arrow : default = 1}

\item{arrow_shape}{Shape of the arrow : "straight" (default), "bhrc", "fhrc", "bhio", "fhio" (give curved throwing shapes)}

\item{label_pos}{Position of label relative to start or end of arrow : "start_down" (default), "start_up", "end_down", "end_up"}
}
\value{
A data.frame
}
\description{
Create a pitch_arrow dataframe. Arrows are represented with geom_segment or geom_curve in the plot_play() function.
Arrows are usually player movements (cuts) or throws on the field.
}
\examples{
pitch_arrow("T1", type = "Throw", x = 5, y = 37, xend = 5, yend = 90, frame = 3)
pitch_arrow("C1", x = 5, y = 37, xend = 5, yend = 90, frame = 3)
}
