% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{pitch_object}
\alias{pitch_object}
\title{Create a pitch_object dataframe. Objects are represented with geom_point in the plot_play() function.
Objects are usually players on the field.}
\usage{
pitch_object(
  label = "",
  show = F,
  object = "Offense",
  x,
  y,
  frame = 1,
  alpha = 0.5,
  label_pos = "down"
)
}
\arguments{
\item{label}{Name for the object (required)}

\item{show}{Show the label in plots : default = F}

\item{object}{Name for the type of object : "Offense" (default), "Defense", "Coach", "Cone", "Disc" are recognised but can be anything.}

\item{x}{Vector of x-coords covered by the object (required)}

\item{y}{Vector of y-coords covered by the object (required)}

\item{frame}{Vector of frames to show the object in : default = 1}

\item{alpha}{Colour opacity of the object : default = 0.5}

\item{label_pos}{Position of label relative to object : "down" (default), "up", "left", "right"}
}
\value{
A data.frame
}
\description{
Create a pitch_object dataframe. Objects are represented with geom_point in the plot_play() function.
Objects are usually players on the field.
}
\examples{
pitch_object("H1", T, x = c(15,15,5), y = c(25,35,37), frame = 1:3)
pitch_object("D1", F, object = "Defense", x = c(13,13,8), y = c(25,35,37), frame = 1:3)
pitch_object("q1", F, object = "Cone", x = c(13), y = c(26), frame = 1:3)
}
