% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{plot_play}
\alias{plot_play}
\title{Plotting function to create final graphic with a pitch, objects and arrows.}
\usage{
plot_play(
  pitch = ggpitch(),
  arrow_list = NULL,
  object_list = NULL,
  static_frame = 1,
  animate = F,
  keep_arrows = F
)
}
\arguments{
\item{pitch}{A pitch ggplot object create by ggpitch() : default = ggpitch()}

\item{arrow_list}{List of pitch_arrow dataframes}

\item{object_list}{List of pitch_object dataframes}

\item{static_frame}{Which frame to show in a static plot : default = 1}

\item{animate}{Show all frames so plot can be animated : F (default), T}

\item{keep_arrows}{Show arrows in all frames and only apply frames to objects : F (default), T}
}
\value{
A ggplot object
}
\description{
Plotting function to create final graphic with a pitch, objects and arrows.
}
\examples{
plot_play(pitch,arrow_list,object_list,static_frame=2)
}
