% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggulti.R
\name{static_objects}
\alias{static_objects}
\title{Create multiple a list of pitch_objects which do not move throughout all frames. Objects are represented with geom_point in the plot_play() function.
Static objects could be cones or players in a queue.}
\usage{
static_objects(
  label,
  show = F,
  object = "Offense",
  x,
  y,
  frame = 1,
  alpha = 0.5,
  label_pos = "down"
)
}
\arguments{
\item{label}{Name for the object (required)}

\item{show}{Show the label in plots : default = F}

\item{object}{Name for the type of object : "Offense" (default), "Defense", "Coach", "Cone", "Disc" are recognised but can be anything.}

\item{x}{Vector of x-coords covered by the object (required)}

\item{y}{Vector of y-coords covered by the object (required)}

\item{frame}{Vector of frames to show the object in : default = 1}

\item{alpha}{Colour opacity of the object : default = 0.5}

\item{label_pos}{Position of label relative to object : "down" (default), "up", "left", "right"}
}
\value{
A data.frame
}
\description{
Create multiple a list of pitch_objects which do not move throughout all frames. Objects are represented with geom_point in the plot_play() function.
Static objects could be cones or players in a queue.
}
\examples{
static_objects(label="Cone", object = "Cone",x = c(10,28),y = 62,alpha = 0.5, frame=1:4),
static_objects(label="queue", x = c(10,28), y = c(58,58,56,56,54,54,52,52), alpha = 0.2, frame=1:4)
}
